// Qanat.Web/src/app/shared/generated/model/external-map-layer-simple-dto.ts
export class ExternalMapLayerSimpleDto { 
    ExternalMapLayerID?: number;
    GeographyID?: number;
    ExternalMapLayerDisplayName?: string | null;
    ExternalMapLayerTypeID?: number;
    ExternalMapLayerTypeDisplayName?: string | null;
    ExternalMapLayerURL?: string | null;
    LayerIsOnByDefault?: boolean;
    IsActive?: boolean;
    ExternalMapLayerDescription?: string | null;
    PopUpField?: string | null;
    MinZoom?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ExternalMapLayerSimpleDtoForm { 
    ExternalMapLayerID?: FormControl<number>;
    GeographyID?: FormControl<number>;
    ExternalMapLayerDisplayName?: FormControl<string>;
    ExternalMapLayerTypeID?: FormControl<number>;
    ExternalMapLayerTypeDisplayName?: FormControl<string>;
    ExternalMapLayerURL?: FormControl<string>;
    LayerIsOnByDefault?: FormControl<boolean>;
    IsActive?: FormControl<boolean>;
    ExternalMapLayerDescription?: FormControl<string>;
    PopUpField?: FormControl<string>;
    MinZoom?: FormControl<number>;
}

export class ExternalMapLayerSimpleDtoFormControls { 
    public static ExternalMapLayerID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static GeographyID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ExternalMapLayerDisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ExternalMapLayerTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ExternalMapLayerTypeDisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ExternalMapLayerURL = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static LayerIsOnByDefault = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static IsActive = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ExternalMapLayerDescription = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static PopUpField = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static MinZoom = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}